refactor: split dashboard sections into components
